Add files via upload
the "signup, sign in, home, about, profile" .tsx file is from the "page" folder under the client/src folder, the "header.tsx" file is under the client/src/components folder, the index.ts is fromserver folder. here we successfully restart our work and also our signup page is connected to the database, so when we fill up the signup page the data will directly goes into the db.